Refactor core Page for reuse
